+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/JavaParserVisitorAdapter.java
@@ -664,13 +664,6 @@ public class JavaParserVisitorAdapter implements JavaParserVisitor {
     }
 
     @Experimental
-    @Deprecated
-    @Override
-    public Object visit(ASTGuardedPattern node, Object data) {
-        return visit((JavaNode) node, data);
-    }
-
-    @Experimental
     @Override
     public Object visit(ASTSwitchGuard node, Object data) {
         return visit((JavaNode) node, data);