+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/JavaParserVisitorDecorator.java
@@ -796,13 +796,6 @@ public class JavaParserVisitorDecorator implements JavaParserControllessVisitor 
     }
 
     @Experimental
-    @Deprecated
-    @Override
-    public Object visit(ASTGuardedPattern node, Object data) {
-        return visitor.visit(node, data);
-    }
-
-    @Experimental
     @Override
     public Object visit(ASTSwitchGuard node, Object data) {
         return visitor.visit(node, data);