+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/ast/KotlinTestingDsl.kt
@@ -26,7 +26,7 @@ enum class JavaVersion : Comparable<JavaVersion> {
     J15,
     J16,
     J17,
-    J18, J18__PREVIEW,
+    J18,
     J19, J19__PREVIEW,
     J20, J20__PREVIEW;